Send messages without a declared interface (ie. bind) correctly
The bind message is perhaps the only example of a message with a
new_id argument and no declared interface. Such new_id arguments are
preceeded with additional interface name and version arguments. Such
messages use the versioned constructor marshalling procedure. Also,
that interface name argument is a string, not an int.